[BugFix: ZSTAC-40686]
Fix delete bridge error

If no l3 network is created based on the l2 network, the bridge device of the l2 on host will not have any ip address. We check the IP address bound to the bridge device before executing the deletion of the bridge device to avoid errors.