<fix>[zstackctl]: a commit added for commit <7e9c2ba225a24c822a98bf452177b2b367c0dd89>
case: After incremental backup is performed on the remote device,
redundant files are deleted. if the value of 'keep-amount'
or 'remote-keep-amount' is more than but less than the twice
of the amount of files on remote, it will delete files incorrectly.

solution: After an incremental backup, the remote files are deleted
only when the number of remote files exceeds the specified amount of retained files

Resolves: ZSTAC-43046

Change-Id: I79707378776d6c71636379756576666a7564796b