<fix>[zstackctl]: fix file loss when backing up database to remote
cause: After incremental backup is performed on the remote device,
redundant files are deleted. if the value of 'keep-amount'
or 'remote-keep-amount' is more than but less than the twice
of the amount of files on remote, it will delete files incorrectly.

solution: After an incremental backup, the remote files are deleted
only when the number of remote files exceeds the specified amount of retained files

fix issue caused by commit <7e9c2ba2>

Resolves: ZSTAC-43046

Change-Id: I6e6561676a756a7a6169616576636a6f72786d65