<fix>[kvmagent]: transfer pvpanic configuration through monitoring
The VM pvpanic configuration was originally reported through the command
reply. However, the frequency of sending GetVmGuestToolsInfoCmd
is too high, the host queue on the MN is easy to be stuck,
so the operation on the host takes a lot of time.

Therefore, The VM pvpanic configuration will be reported
through Prometheus to relieve the pressure on the host queue.

Resolves: ZSTAC-49036
Releted: ZSTAC-47374

Change-Id: I70797161617165736864786e71766a6768706a66