<fix>[kvmagent]: check disk state after detach temp disk
When the VM executes the detach disk command during startup,
libvirt will report that the execution is successful,
but the disk is still in the mount list of the VM domain.

In fact, libvirt/qemu just sends the detach disk command
to the guest. Without waiting for the guest to return,
it directly exits and reports success result.

In the Kvm-agent, we need to ensure that all the disks
of guest tools have been detached, so that the subsequent
hot migration can proceed smoothly. Therefore,
we need to check the status of the disk again
after sending the detach disk command.

Resolves: ZSTAC-48137

Change-Id: I656c67706a7270627065796c6566697a6172786e