<fix>[ansible]: fix libvirt can not downgrade automatic on ns10
If the zstack upgrade from 4.4.9(libvirt 6.2.0) to 4.5.0(libvirt 6.0.0),
the libvirt can not downgrade to 6.0.0 automatic, therefore, downgrade
libvirt manually after install required rpms. The downgrade condition
is that the system is x86_64_ns10, and the libvirt version of the host
is inconsistent with the libvirt version of repo.
Interestingly, upgrading from 4.2.9 to 4.5.0 did not have this problem,
the libvirt can downgrade during upgrade qemu rpms, but need time to
research.

Resolves: ZSTAC-50272

Change-Id: I7669796e6276616662676a696c676d6967657665