<fix>[cephbackupstorage]: Fix failed deploy cephps
Since 4.5.11, qemu-kvm-ev renamed to qemu-kvm in c76 and c79, however,
c74 do not apply the change. If using c74 as ceph host, and using
c76/c79 as mn, cephbs will desire to install qemu-kvm-ev rpm, which is
not existed in c76/c79 repo.
Therefore, using qemu-kvm rpm in the case.

Resolves: ZSTAC-52868

Change-Id: I7175696f6c68796e71676364727a756972696c69