<feature>[zstackbuild]: Upgrade ansible to 4.10.0 & optimize code style
In this patch, ansible is upgraded to 4.10.0, which is the last version
that can support python2.

For this, additional dependencies need to be added, as listed below:
- ansible-4.10.0.tar.gz
- ansible-core-2.11.12.tar.gz
- packaging-20.9.tar.gz
- pathlib-1.0.1.tar.gz
- pyparsing-2.4.7.tar.gz
- resolvelib-0.5.5-py2.py3-none-any.whl
- scandir-1.10.0.tar.gz
- setuptools-39.2.0-py2.py3-none-any.whl

In addition, due to a major upgrade of ansible, the old ansible.runner
call was no longer works, and the ansible-runner project requres
python3, therefore, the ansible.runner call is replaced by calling
ansible api directly.

Also, some of the formatting has been optimized to comply with the pep8
format. It's recommended to clearly import the required modules rather
that import *

Resolves: ZSTAC-52322

Change-Id: I736261637270796867786e69717175737a636271