<fix>[zstackbuild]: Fix failed to connect kylin host
On ansible 1.9.6, the distribution value returned by gather facts was
'Kylin', and distribution_version/distribution_major_version were '10'.
But on ansible 4.10.0, the information returned is 'Kylin Linux Advanced
Server' and 'V10', and this inconsistency causes erroes in the
associated logic(such as connect host).
Therefore, modify the ansible gather facts logic, add the processing of
kylin ditro so that it's consistent with the returns of ansible 1.9.6.

Resolves: ZSTAC-56031

Change-Id: I7175786279736368716270746d75676e736b6d71