<fix>[ansible]:Clean up iptables rules related to kvmagent ports
This commit removes the KVM Agent port rules in the iptables
configuration to ensure that the port is cleared and functioning properly.
This check is important to ensure that the KVM virtualization environment is
properly configured and ready for use.

If the kvmagent port is blocked and the echo fails, the kvmagent
will be restarted and the command will be reissued.
However, due to the kvmagent port being blocked, the command will
fail to be issued.
The HA fencer is unable to update the heartbeat, and during the HA checker
check, the heartbeat will not be refreshed, potentially resulting in
split-brain VMs.

Resolves: ZSTAC-56336

Change-Id: I656f6b79646268676d7476736770666f68736474