<fix>[zstackctl]: Fix failed check log size
Fix failed check log size, recently, there was a new item:
/etc/yum.repos.d add into log collector. But the method:
`run_remote_command` will detect if there is 'yum' characters in the
command , and add commands with `echo` `grep` to detect releasever,
which will cause the result returned have `c79\nc79\n` characters,
resulting in a failure to parse the result, and the log size check exits
abnormally.
Therefore, and a new method param: `with_match` for method
`add_command_action`, if the param is True, then detect whether `yum`
was included in the command and do the action as above, if not then do
not detect `yum` chatacters. For compatibility, this parameter defaults
to True.

Resolves: ZSTAC-56669

Change-Id: I6e65747869686c626d6d646c6d786d6f74796c61