<fix>[vm]: Fix failed live migrate VM due to cpu check incorrect
A recent update of live migration is to use `MIGRATEABLE` param to get
the domain xml, where the CPU part omits some configuration items to get
better compatibility, such as `check=full`. This logic works fine for
libvirt-6.0.0-560 and later. However, for previous 6.0.0 versions of
libvirt, it checks CPU configuration changes which can cause failed to
take live migration.
Therefore, check the libvirt version info, when the version is 6.0.0 and
release < 560, still use the old way to get the domain xml, in other
cases through the `MIGRATEABLE` param to get the domain xml.

Resolves: ZSTAC-61974

Change-Id: I7a786d6c7569766d756d636176786b646d686c63