<fix>[host]: Fix failed find suitable qemu-kvm rpm package
During the updateClusterOS execution, the qemu-kvm replacement is done
before updating the specify rpm, and the yum repo cache is not cleaned
at this time, which will search qemu-kvm package in the old repo and
cause failed to replace qemu-kvm package.
Therefore, clean the yum repo cache before updating the rpm to avoid
using an unexpected repo.

Resolves: ZSTAC-60240,ZSTAC-61881

Change-Id: I74687469686d716d696a736d67666775616f6c6b