<fix>[vm]: detach volume should wait for result if already under detaching
1.New version qemu will raise exception when try to detach (device_del in qemu)
a volume which is already in the process of unplugging, so just ignore the error
but wait for the volume detached.
2.Add a function called "ignore_exception(exception_type, message=None)"
that ignores a specific type of exception.
The "exception_type" parameter indicates the type of exception to ignore.
If the "message" parameter is None,
the exception will be ignored and a log message will be printed.
If the "message" parameter is not None,
the exception will be raised if the message string is contained within the exception.

Resolves: ZSTAC-62997

Change-Id: I6d6b787866757a7061656b716b73766a73677478