<fix>[zstacklib]: fix nfs cache inconsistency during create directory
For nfs fs, the attributes of the files are stored in the attribute
cache, os it may happend that host1 deletes the file, but host2 decides
that the file still exists because the cache is not refreshed.
Therefore, access the file/directory to trigger a cache refresh before
determine if the file exists.

Resolves: ZSV-4288

Change-Id: I776775717669647676716174776c786567767171