<fix>[vm]: Fix failed to take snapshot after upgrade to qemu6.2.0
ZStack uses the file type to drive sblk devices in libvirt. However,
qemu6.2.0 do not support the usage, therefore need to detect if the disk
is a sblk device and set it to block type when create snapshot. Also,
during the live migration, detect whether the snapshot chain has sblk
device but with file type, correct it to block type in dest xml before
do live migrate.

Resolves: ZSTAC-66696

Change-Id: I716673657068656a7a667972796d6970656c6b67