<fix>[ha]: wait ceph hb object to be created during ha peer checking
Sometimes ceph heartbeat object may not be found which will
failed directly when use ioctx.stat(). But checker should
only wait for heartbeat count not changed or heartbeat object
not created as the same result that peer host is not connected
to the ceph storage. Instead of a straight failure, wait for
storage check timeout to confirm that we could not get the
length from ceph heartbeat object.

Resolves: ZSTAC-65952

Change-Id: I68676b73706a686476726863797a79736f7a6b78
Signed-off-by: AlanJager <ye.zou@zstack.io>